+++ src/grammar-support/log-manager.js
@@ -1,3 +1,5 @@
+const TagList = require('./tag-list');
+
 class BranchList {
   constructor() {
     /**
@@ -67,28 +69,6 @@ class Branch {
    */
   getCommitCount() {
     return this._commits;
-  }
-}
-
-class TagList {
-  constructor() {
-    /**
-     * @type {Set<string>}
-     * @private
-     */
-    this._tags = new Set();
-  }
-
-  /**
-   * @param {string} tag
-   * @throws {Error}
-   */
-  add(tag) {
-    if (this._tags.has(tag)) {
-      throw new Error('Tag already exists: ' + tag);
-    }
-
-    this._tags.add(tag);
   }
 }
