+++ src/grammar-support/tag-list.js
@@ -1,0 +1,23 @@
+class TagList {
+  constructor() {
+    /**
+     * @type {Set<string>}
+     * @private
+     */
+    this._tags = new Set();
+  }
+
+  /**
+   * @param {string} tag
+   * @throws {Error}
+   */
+  add(tag) {
+    if (this._tags.has(tag)) {
+      throw new Error('Tag already exists: ' + tag);
+    }
+
+    this._tags.add(tag);
+  }
+}
+
+module.exports = TagList;
